sys-devel/perl-5.8.0-r7: Error found in 5.8 ebuild for non threaded perls
PLEASE READ. There was a problem introduced for non threaded perl 5.8's
- a correction introduced earlier for threaded perls was adding
-thread to the name of your arch. THIS CAUSES PROBLEMS FOR SOME
MODULES. For instance, the DBI module checks to see what the name of
your Config.pm has listed for $Config{archname} - if it had thread in
it *anywhere*, it assumed you had threading enabled, but since your
perl was failing to use threads, would fail. This will version of the
ebuild will correct that. You will need to re-emerge modules installed
with perl 5.8. A tool for this is forthcoming.