sys-devel/perl-5.8.0-r8: syntax updates
ebuild: put a sleep where we had one in the else block so that there is
a pause before emerging (notice regarding threads). Added a sleep at
the tail end of the ebuild so that there is a pause for folks doing an
emerge -u world so that they at least have an opportunity to see the
notice.

libperl_rebuilder: removed the unmerge section, that's really overkill
for our needs and only complicates things. Added new syntax to check
for files installed into /usr/lib/perl* that weren't picked up
previously. Changed the name of the log file to perl-update.log
(makes more sense ;) ) Added mask check earlier in the sanity phase.