sys-devel/perl-5.8.0-r9: Fix bugs #13886 and #13920
Major typo corrections that affected the placement of scripts and the
man pages. There was a typo introduced into the 5.8 ebuilds early on
that was not caught until now that left the location of the man pages
blank and that misset the /usr call for scripts placement from perl
modules. This was caught in bugs 13886 and 13920

Bug: https://bugs.gentoo.org/13886
Bug: https://bugs.gentoo.org/13920