dev-lang/perl-5.8.4-r1: Many fixes
Move make test to src_test, but still call it if the maketest FEATURE
is not enabled. If maketest ever becomes the default, this can be
reconsidered. The NDBM-GDBM patch in bug #52660 is in here and gdbm is
allowed to provide ndbm, as it works for me with either USE=berkdb and
have db1 installed, or USE=gdbm in which case gdbm will come in. The
man page fixes from bug #58620 are in.

src_configure has been split out of src_compile in preparation for
making compiles more easily resumable. An issue genone brought up where
hardlinks aren't making it through portage's staging image has been
addressed by making /usr/bin/perl a symlink to perl5.8.4 instead of a
hardlink. The same issue was faced with suidperl, which has been
removed completely to preemptively avoid security issues. sperl${PV}
is gone to, in favor of the perl recommendation to use sudo instead
of having setuid perl executables. I hope this will help security, but
it can be restored if there is enough demand.

Bug: https://bugs.gentoo.org/52660
Bug: https://bugs.gentoo.org/58620